+++ noodles-bcf/src/reader/record.rs
@@ -261,13 +261,10 @@ pub(crate) mod tests {
     #[test]
     fn test_read_record() -> Result<(), Box<dyn std::error::Error>> {
         use noodles_vcf::{
-            header::info::Key as InfoFieldKey,
+            header::{format::Key as GenotypeFieldKey, info::Key as InfoFieldKey},
             record::{
                 genotypes::{
-                    genotype::{
-                        field::{Key as GenotypeFieldKey, Value as GenotypeFieldValue},
-                        Field as GenotypeField,
-                    },
+                    genotype::{field::Value as GenotypeFieldValue, Field as GenotypeField},
                     Genotype, Keys,
                 },
                 info::{field::Value as InfoFieldValue, Field as InfoField},