+++ noodles-vcf/src/record/genotypes/genotype/field/value.rs
@@ -145,7 +145,10 @@ impl Value {
     /// # Examples
     ///
     /// ```
-    /// use noodles_vcf::record::genotypes::genotype::field::{Key, Value};
+    /// use noodles_vcf::{
+    ///     header::format::Key,
+    ///     record::genotypes::genotype::field::Value,
+    /// };
     ///
     /// assert_eq!(
     ///     Value::from_str_key("13", &Key::ConditionalGenotypeQuality),